Rename InterestData file to start with lower case
